# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: AionUI WebUI API
  version: 1.8.1
  description: |
    REST API for AionUI WebUI server providing authentication, conversation management,
    and AI agent interaction capabilities.

    ## Authentication
    AionUI supports two authentication methods:
    - **Local**: Username/password authentication with JWT tokens
    - **OIDC/SSO**: Enterprise single sign-on via EntraID (Azure AD) or other OIDC providers

    All protected endpoints require JWT authentication via:
    - Cookie: `aion_session` (access token, 15 min expiry)
    - Header: `Authorization: Bearer <token>`
    
    Refresh tokens (7 day expiry) are issued via httpOnly `aionui-refresh` cookie.
    Use `/api/auth/refresh` to rotate tokens without re-authentication.

    ## Authorization (RBAC)
    Users have one of three roles:
    - **admin**: Full system access including user management
    - **user**: Standard access to conversations and AI features
    - **viewer**: Read-only access

    ## Rate Limiting
    - Login endpoints: 5 requests per minute
    - API endpoints: 100 requests per minute
    - Authenticated actions: 30 requests per minute

  contact:
    name: Jason Matherly
    email: jason@mathrly.net
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:25808
    description: Local development server
  - url: https://{host}:{port}
    description: Production server
    variables:
      host:
...
      port:
...
paths:
  /login:
    post:
      tags: [Authentication]
      summary: User login
      description: Authenticate user with username and password
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          headers:
            Set-Cookie:
              schema:
                type: string
              description: Session cookie (aion_session)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
  /logout:
    post:
      tags: [Authentication]
      summary: User logout
      description: Invalidate current session and clear cookies
      operationId: logout
      security:
        - cookieAuth: []
        - bearerAuth: []
      responses:
        "200":
          description: Logout successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Logged out successfully"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/auth/status:
    get:
      tags: [Authentication]
      summary: Get authentication status
      description: Check if system needs initial setup and user count
      operationId: getAuthStatus
      responses:
        "200":
          description: Authentication status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthStatusResponse"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/auth/user:
    get:
      tags: [Authentication]
      summary: Get current user
      description: Retrieve current authenticated user information
      operationId: getCurrentUser
      security:
        - cookieAuth: []
        - bearerAuth: []
      responses:
        "200":
          description: User information
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/auth/change-password:
    post:
      tags: [Authentication]
      summary: Change password
      description: Change the current user's password
      operationId: changePassword
      security:
        - cookieAuth: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "200":
          description: Password changed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Password changed successfully"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/auth/refresh:
    post:
      tags: [Authentication]
      summary: Refresh access token
      description: |
        Rotate refresh token and issue new access + refresh tokens.
        Uses httpOnly `aionui-refresh` cookie for authentication.
      operationId: refreshToken
      responses:
        "200":
          description: Tokens refreshed successfully
          headers:
            Set-Cookie:
              schema:
                type: string
              description: New httpOnly aionui-refresh cookie
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefreshTokenResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/auth/oidc/login:
    get:
      tags: [Authentication]
      summary: Initiate OIDC login
      description: |
        Redirects to the configured OIDC identity provider (EntraID/Azure AD)
        for authentication. Available when OIDC is enabled.
      operationId: oidcLogin
      responses:
        "302":
          description: Redirect to identity provider
          headers:
            Location:
              schema:
                type: string
              description: OIDC authorization endpoint URL
        "503":
          description: OIDC not configured
  /api/auth/oidc/callback:
    get:
      tags: [Authentication]
      summary: OIDC callback handler
      description: |
        Handles the callback from OIDC provider. Exchanges authorization code
        for tokens, performs JIT (just-in-time) user provisioning if needed,
        and establishes session.
      operationId: oidcCallback
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
          description: Authorization code from OIDC provider
        - name: state
          in: query
          required: true
          schema:
            type: string
          description: State parameter for CSRF protection
        - name: session_state
          in: query
          schema:
            type: string
          description: Optional session state from provider
      responses:
        "302":
          description: Redirect to application with session established
          headers:
            Location:
              schema:
                type: string
              description: Application URL (typically /)
            Set-Cookie:
              schema:
                type: string
              description: Session cookies (aion_session, aionui-refresh)
        "401":
          description: Authentication failed
          content:
            text/html:
              schema:
                type: string
  /api/auth/qr-login:
    post:
      tags: [Authentication]
      summary: QR code login
      description: Authenticate using QR code token (for mobile/desktop pairing)
      operationId: qrLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [qrToken]
              properties:
                qrToken:
                  type: string
                  description: QR token from scanned code
      responses:
        "200":
          description: QR login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/ws-token:
    get:
      tags: [WebSocket]
      summary: Get WebSocket token
      description: Get token for WebSocket connection (reuses main session token)
      operationId: getWsToken
      security:
        - cookieAuth: []
        - bearerAuth: []
      responses:
        "200":
          description: WebSocket token
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  wsToken:
                    type: string
                  expiresIn:
                    type: integer
                    description: Token expiry in milliseconds
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token (15 min expiry)
    cookieAuth:
      type: apiKey
      in: cookie
      name: aion_session
      description: Session cookie containing JWT access token
    adminAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token with admin role required
  schemas:
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 1
          example: admin
        password:
          type: string
          minLength: 8
          format: password
    LoginResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
          description: JWT token
    User:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        role:
          type: string
          enum: [admin, user, viewer]
          description: User's role determining access level
        authMethod:
          type: string
          enum: [local, oidc]
          description: Authentication method used
        displayName:
          type: string
          nullable: true
          description: User's display name (from OIDC or profile)
        email:
          type: string
          format: email
          nullable: true
          description: User's email address
    AuthStatusResponse:
      type: object
      properties:
        success:
          type: boolean
        needsSetup:
          type: boolean
          description: True if no users exist
        userCount:
          type: integer
        isAuthenticated:
          type: boolean
        oidcEnabled:
          type: boolean
          description: True if OIDC/SSO authentication is configured
    UserResponse:
      type: object
      properties:
        success:
          type: boolean
        user:
          $ref: "#/components/schemas/User"
...
    RefreshTokenResponse:
      type: object
      properties:
        success:
          type: boolean
        token:
          type: string
          description: New access token (also set in aion_session cookie)
        expiresIn:
          type: integer
          description: Access token expiry in seconds (900 = 15 min)
    ChangePasswordRequest:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
          format: password
        newPassword:
          type: string
          format: password
          minLength: 8
          description: Must meet security requirements
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
        message:
          type: string
        code:
          type: string
  responses:
    Unauthorized:
      description: Authentication required or invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            success: false
            message: "Invalid username or password"
...
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    TooManyRequests:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            success: false
            error: "Too many requests, please try again later"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            success: false
            message: "Internal server error"